+++ tests/host/conanfile.py
@@ -7,7 +7,7 @@ class RevenantHostTests(ConanFile):
     python_requires = "ios6-base/1.0@ios6/stable"
 
     def requirements(self):
-        self.requires("icu/[*]@ios6/stable")
+        self.requires("icu/[*]@revenant/stable")
 
     def generate(self):
         self.python_requires["ios6-base"].module.DependencyEnv(self).generate()